Move Flume message protocol to the flume package
